requirements.txt
asgiref==3.6.0
Django==4.1.6
django-jazzmin==2.6.0
django-phone-field==1.8.1
path==16.6.0
Pillow==9.4.0
sqlparse==0.4.3
tzdata==2022.7
